# KAFKA-13512: topicIdsToNames and topicNamesToIds allocate unnecessary maps