@dajac 2021-12-07
The MetadataSnapshot is immutable so this seems fine to me as well.